58d9802 clang-format .c sources (Jonas Schnelli)
6c597c2 clang-format headers (Jonas Schnelli)
37af5a8 Header cleanup (Jonas Schnelli)

Tree-SHA512: aebb5365cccce36b10b53c5647e61ca95cd1cae7e12eda71095385b671133c2f78b4b5afe26be71a7c082ec2b8319cf53aad05d5639bff3754ab40c89d434930

+++ test/tx_tests.c
@@ -781,7 +781,7 @@ void test_script_parse()
         btc_script_get_ops(script, vec);
         enum btc_tx_out_type type = btc_script_classify_ops(vec);
 
-        assert(type == test->type);
+        assert((int)type == test->type);
         assert(vec->len == (size_t)test->opcodes);
 
         vector_free(vec, true);
